refactor: Consolidate parser files and remove unnecessary indirection
- Move RouteParserFacade logic directly into RoutePatternParser
- Delete ImprovedRoutePatternParser.cs file (confusing name)
- Remove unused TokenizePattern method (dead code from old regex parser)
- Remove unnecessary Segments namespace from using statements

This simplifies the architecture by removing an unnecessary layer of
indirection. RoutePatternParser is now the single public API that
directly uses the internal RouteParser and RouteCompiler.
