+++ Source/TimeWarp.Nuru/GlobalUsings.cs
@@ -16,5 +16,4 @@ global using TimeWarp.Nuru.CommandResolver;
 global using TimeWarp.Nuru.Endpoints;
 global using TimeWarp.Nuru.Help;
 global using TimeWarp.Nuru.Parsing;
-global using TimeWarp.Nuru.Parsing.Segments;
 global using TimeWarp.Nuru.TypeConversion;
